+++ plugins/main/public/components/overview/vulnerabilities/dashboards/inventory/inventory.tsx
@@ -228,10 +228,12 @@ const InventoryVulsComponent = () => {
                 filters={excludeUnderEvaluationFilter(filters)}
                 fixedFilters={fixedFilters}
                 postFixedFilters={[
-                  <VulsEvaluationFilter
-                    value={underEvaluation}
-                    setValue={handleFilterChange}
-                  />,
+                  () => (
+                    <VulsEvaluationFilter
+                      value={underEvaluation}
+                      setValue={handleFilterChange}
+                    />
+                  ),
                 ]}
                 showDatePicker={false}
                 showQueryInput={true}