+++ CHANGELOG.md
@@ -8,8 +8,9 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 - Support for Wazuh 4.13.0
 - Added `It Hygiene` application [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
-- Added hardware and system information to agent overview [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
-- Added `fim.pattern` and `system_inventory.pattern` settings [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
+- Added hardware and system information to the agent overview [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
+- Added settings to manage the FIM and IT Hygiene inventories data [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
+- Add persistence for selected columns and page size in data grid settings [#7379](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7379)
 
 ### Removed
 
@@ -17,7 +18,13 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Removed `GET /api/syscollector` API endpoint [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
 - Removed inventory data report and `POST /reports/agents/{agentID}/inventory` API endpoint [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
 
-## Wazuh v4.12.0 - OpenSearch Dashboards 2.19.1 - Revision 00
+## Wazuh v4.12.1 - OpenSearch Dashboards 2.19.1 - Revision 00
+
+### Added
+
+- Support for Wazuh 4.12.1
+
+## Wazuh v4.12.0 - OpenSearch Dashboards 2.19.1 - Revision 01
 
 ### Added
 
@@ -39,6 +46,10 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Fixed CSV export not filtering by timerange [#7304](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7304)
 - Fixed agent view not showing the latest agent state [#7336](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7336)
 - Fixed saved queries not displaying in the search bar [#7377](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7377)
+
+### Removed
+
+- Removed revision number from about page [#7390](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7390)
 
 ## Wazuh v4.11.2 - OpenSearch Dashboards 2.16.0 - Revision 02